requirements.txt
# Core Dependencies for Architectural Style Classification Pipeline
# Install with: pip install -r requirements.txt

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
Pillow>=9.5.0
PyYAML>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars
tqdm>=4.65.0

# Experiment Tracking
mlflow>=2.8.0

# System Metrics Monitoring (optional but recommended for MLflow)
psutil>=5.9.0              # CPU, memory, disk, network monitoring
nvidia-ml-py>=12.535.77    # GPU monitoring (NVIDIA GPUs only)

# Model Explainability
# Note: gradcam implementation is custom, no external dependency needed

# Optional: Hyperparameter Tuning
scikit-learn>=1.3.0        # For genetic algorithms and utilities

# Optional: Development Tools
# ipython>=8.12.0          # Better REPL
# jupyter>=1.0.0           # Notebook support
# pytest>=7.3.0            # Testing framework
